Move Services into internal/codeintel.
